Problem: Lack of robustness around transactions
Solution: retry transaction submission.

When transactions are not mined during the monitoring period, the
transaction should be re-submitted with a higher gas price to increase
the chance of it getting mined.

Remove the logger attribute of the mediator and use a module level
logger consistent with other parts of the code base.

- Remove commit await.
- Manual nonce management (just for commit).
- Monitor tx periodically (every n secs/blocks).
- If tx not mined, replace tx with higher gas price.
- geth: enable miner rpc API.
- Test failed tx creation retry.
- Test failed mining retry.
- Use context for knex logs.